+++ src/OplogMonitor.h
@@ -1,0 +1,37 @@
+/*
+ * OplogMonitor.h
+ *
+ * This source file is part of the FoundationDB open source project
+ *
+ * Copyright 2013-2018 Apple Inc. and the FoundationDB project authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+#ifndef _OPLOG_MONITOR_H
+#define _OPLOG_MONITOR_H
+#pragma once
+
+#include "bindings/flow/fdb_flow.h"
+#include "bindings/flow/DirectorySubspace.h"
+#include "flow/flow.h"
+#include "DocLayer.h"
+#include "ExtStructs.h"
+
+
+void oplogSendTimestamp(PromiseStream<double> times, double ts);
+void oplogRunUpdateScanner(FutureStream<double> times, Reference<DocumentLayer> docLayer, Reference<ExtChangeStream> changeStream);
+void oplogRunStreamWatcher(Reference<DocumentLayer> docLayer, FutureStream<double> times);
+void oplogRunTimestampWatcher(PromiseStream<double> times, Reference<DocumentLayer> docLayer);
+void oplogMonitor(Reference<DocumentLayer> docLayer, double logsDeletionOffset);
+
+#endif /* _OPLOG_MONITOR_H */
